[Silver II] Title: 키트 분배하기, Time: 340 ms, Memory: 33512 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/24938. 키트 분배하기/키트 분배하기.java
@@ -6,16 +6,16 @@ public class Main {
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		final int N = Integer.parseInt(br.readLine()); // numbers of room
-		int[] arr = new int[N];
+		long[] arr = new long[N];
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		int sum = 0;
+		long sum = 0;
 		for (int i = 0; i < N; i++) {
-			arr[i] = Integer.parseInt(st.nextToken());
+			arr[i] = Long.parseLong(st.nextToken());
 			sum += arr[i];
 		}
-		final int goal = sum / N;
-		int chaos = 0;
-		int rest = 0;
+		final long goal = sum / N;
+		long chaos = 0;
+		long rest = 0;
 		for (int i = 0; i < N; i++) {
 			chaos += Math.abs(rest);
 			rest += arr[i] - goal;
